+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - A service with a single path parameter now has path parameters of `(T,)`, not `OnePathParameter`.
   - A service with multiple path parameters now has a tuple of path parameters, not `ManyPathParameters`.
 - Moved Status Code constants to inside [StatusCode](https://docs.rs/picoserve/latest/picoserve/response/status/struct.StatusCode.html).
+- [FromRequest](https://docs.rs/picoserve/latest/picoserve/extract/trait.FromRequest.html) and [FromRequestParts](https://docs.rs/picoserve/latest/picoserve/extract/trait.FromRequestParts.html) are now generic over the lifetime of the request, allowing them to borrow from the request.
+
+### Added
+
+- Added [from_request](https://docs.rs/picoserve/latest/picoserve/macro.from_request.html) and [from_request_parts](https://docs.rs/picoserve/latest/picoserve/macro.from_request.html) as convenience for [PathRouters](https://docs.rs/picoserve/latest/picoserve/routing/trait.PathRouter.html), and added [RequestHandlerServices](https://docs.rs/picoserve/latest/picoserve/routing/trait.RequestHandlerService.html) which borrow from Requests, which is now permitted.
 
 ## [0.9.1] - 2024-02-12